Move concepts to its own package
